requirements.txt
Flask==3.1.0
PyNaCl==1.5.0
selenium==4.26.1
undetected_chromedriver==3.5.5
